Linux: use generic net_tstamp API for more accurate time stamping, with and without HW support
The approach is very similar to the looped multicast trick: send time stamps
are handled when the sent packet comes back via the socket's error queue.
This works with and without hardware support in the NIC. With that hardware
support the NIC itself generates the time stamps, otherwise the Linux stack
does it close to sending/receiving the packet.